August 6:     10:30 pm - 11:10 pm (35 min): Tried to make names for the AI. Developed getKingdomName and setKingdomName in Participants. I tried setting the participant names when the participants were created, but createNewName needs there to be more than 0 commanders in the list of commnaders or else the list is null (can't return list of commanders sie if it's null). Had to move the name generation stuff to the end of generateNewContinent, works now. Have developed framework for working with AI and probability of actions and stuff (Read below)
For the AI, have an array like this: int AIActionProbability = [how many actions there are] = {1, 2, 3, 4, 5}

The difference between each element's value affects the probability. For example, since the difference between each element is 1, they will all have an equally likely chance of happening. However, if the array was like this {1, 2, 8, 9, 10}, the differences are the same across the board except for between indices 1 and 2. This means that all other values have the same chance of occurring, but the action that happens between said indices has a greater chance of happening. Applying this, it would be something like: if (randomNumberGenerated >= AIActionProbability[0] && <= AIActionProbability[1]), do some particiular action.